sdk/py/pyproject.toml
[build-system]
requires = ["uv_build>=0.8.14,<0.10.0"]
build-backend = "uv_build"

[project]
name = "petcode"
version = "1.0.0"
description = "PetCode SDK for Python"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "seerbp-petcode-protocolbuffers-python==33.2.0.1.20251225073453+b0513d4afebb",
]

[tool.uv.build-backend]
module-root = "./"
module-name = "petcode"

[[tool.uv.index]]
url = "https://buf.build/gen/python"


[tool.pyright]
pythonVersion = "3.10"
pythonPlatform = "All"
typeCheckingMode = "standard"
reportShadowedImports = false
disableBytesTypePromotions = true

[dependency-groups]
dev = [
    "ruff>=0.13.0",
    "pytest>=9.0.1",
    "tomli>=2.3.0",
]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = [
    "-v",
    "-s",
    "--tb=short",
    "--durations=10",
    "-W", "ignore::DeprecationWarning",
]
minversion = "7.0"
norecursedirs = [".git", ".tox", "dist", "build", "*.egg", ".venv"]
markers = [
    "slow: 标记慢速测试",
    "integration: 标记集成测试",
    "unit: 标记单元测试",
]
